auth, users controllers and middlewares are up to date

+++ controllers/userController.js
@@ -1,8 +1,8 @@
-'use strict';
+"use strict";
 
-const { User, Store, Role, Permission, Contact } = require('../models');
-const logger = require('../config/logger');
-const { paginate } = require('../utils/pagination');
+const { User, Store, Role, Permission, Media } = require("../models");
+const logger = require("../config/logger");
+const { paginate } = require("../utils/pagination");
 
 exports.getUsers = async (req, res, next) => {
   try {
@@ -10,7 +10,15 @@ exports.getUsers = async (req, res, next) => {
     const limit = parseInt(req.query.limit) || 10;
 
     // You can pass additional filters or sorting options if needed
-    const paginatedUsers = await paginate(User, page, limit, {}, [['createdAt', 'DESC']]);
+    const paginatedUsers = await paginate({
+      model: User,
+      page,
+      limit,
+      where: {},
+      options: {
+        order: [["createdAt", "DESC"]],
+      },
+    });
 
     res.status(200).json(paginatedUsers);
   } catch (error) {
@@ -23,18 +31,49 @@ exports.getUserById = async (req, res, next) => {
   try {
     const userId = req.params.id;
 
-    // Fetch the user by ID, including stores, roles, and permissions
+    // Check if the user is accessing their own profile or is an admin
+    const isAdmin = req.user.roles.includes("Admin"); // Check if user has 'Admin' role
+
+    // check if the user is an admin or the user themselves
+    if (!isAdmin && req.user.id !== Number(userId)) {
+      return next({
+        statusCode: 403,
+        message: "You do not have permission to perform this action!",
+      });
+    }
+
+    // Fetch the user by ID, including related data
     const user = await User.findByPk(userId, {
       include: [
-        { model: Store, attributes: ['store_name'] },
-        { model: Role, attributes: ['name'] },
-        { model: Permission, attributes: ['name'] },
-        { model: Contact },
+        {
+          model: Store,
+          as: "Stores",
+          through: { attributes: [] }, // Exclude the UserStore join table
+        },
+        {
+          model: Role,
+          as: "Roles",
+          attributes: ["name", "description"],
+          through: { attributes: [] }, // Exclude the UserRole join table
+          include: [
+            {
+              model: Permission,
+              as: "Permissions",
+              attributes: ["name", "description"],
+              through: { attributes: [] }, // Exclude the RolePermission join table
+            },
+          ],
+        },
+        {
+          model: Media,
+          as: "Media",
+          through: { attributes: [] }, // Exclude the UserMedia join table
+        },
       ],
     });
 
     if (!user) {
-      return next({ statusCode: 404, message: 'User not found' });
+      return next({ statusCode: 404, message: "User not found" });
     }
 
     return res.status(200).json(user);
@@ -42,7 +81,7 @@ exports.getUserById = async (req, res, next) => {
     logger.error(`Error fetching User data: ${error.message}`);
     next(error);
   }
-}
+};
 
 exports.updateUser = async (req, res, next) => {
   try {
@@ -52,12 +91,18 @@ exports.updateUser = async (req, res, next) => {
     const user = await User.findByPk(userId);
 
     if (!user) {
-      return next({ statusCode: 404, message: 'User not found' });
+      return next({ statusCode: 404, message: "User not found" });
     }
 
-    // only admins or the user themselves can update their profile
-    if (req.user.role !== 'admin' && Number(userId) !== req.user.id) {
-      return next({ statusCode: 403, message: 'You do not have permission to perform this action!' });
+    // Check if the user is accessing their own profile or is an admin
+    const isAdmin = req.user.roles.includes("Admin"); // Check if user has 'Admin' role
+
+    // check if the user is an admin or the user themselves
+    if (!isAdmin && req.user.id !== Number(userId)) {
+      return next({
+        statusCode: 403,
+        message: "You do not have permission to perform this action!",
+      });
     }
 
     // If password is provided, it will trigger the beforeUpdate hook to hash it
@@ -73,26 +118,20 @@ exports.updateUser = async (req, res, next) => {
     logger.error(`Error updating User: ${error.message}`);
     next(error);
   }
-}
+};
 
 exports.deleteUser = async (req, res, next) => {
   try {
-    const { id } = req.params;
+    const userId = req.params.id;
 
-    if (!id) {
-      return next({ statusCode: 400, message: 'User ID is required!' });
-    }
-
-    // only admins or the user themselves can delete their profile
-    if (req.user.role !== 'admin' && Number(id) !== req.user.id) {
-      return next({ statusCode: 403, message: 'You do not have permission to perform this action!' });
+    if (!userId) {
+      return next({ statusCode: 400, message: "User ID is required!" });
     }
 
     // Perform a soft delete by setting 'deletedAt' instead of removing the record
-    const deletedUser = await User.destroy({ where: { id } });
-
+    const deletedUser = await User.destroy({ where: { id: userId } });
     if (!deletedUser) {
-      return next({ statusCode: 404, message: 'User not found!' });
+      return next({ statusCode: 404, message: "User not found!" });
     }
 
     res.status(204).send();
@@ -100,27 +139,25 @@ exports.deleteUser = async (req, res, next) => {
     logger.error(`Error deleting User: ${error.message}`);
     next(error);
   }
-}
+};
 
 exports.restoreUser = async (req, res, next) => {
   try {
-    const { id } = req.params;
+    const userId = req.params.id;
 
-    if (!id) {
-      return next({ statusCode: 400, message: 'User ID is required!' });
+    if (!userId) {
+      return next({ statusCode: 400, message: "User ID is required!" });
     }
 
-    const user = await User.restore({ where: { id } });
+    const user = await User.restore({ where: { id: userId } });
 
     if (!user) {
-      return next({ statusCode: 404, message: 'User not found!' });
+      return next({ statusCode: 404, message: "User not found!" });
     }
 
-    res.status(200).json({ message: `User with ID ${id} has been restored` });
+    res.status(200).json({ message: `User with ID ${userId} has been restored` });
   } catch (error) {
     logger.error(`Error restoring User: ${error.message}`);
     next(error);
   }
-}
-
-
+};
